+++ src/Element/Concerns/HasCustomFields.php
@@ -23,6 +23,12 @@ use yii\base\InvalidConfigException;
  * This trait contains all logic related to getting, setting, normalizing,
  * and tracking changes to custom field values on elements.
  *
+ * @property array $serializedFieldValues Array of the element’s serialized custom field values, indexed by their handles
+ * @property array $fieldValues The element’s normalized custom field values, indexed by their handles
+ * @property string $fieldContext The field context this element’s content uses
+ * @property FieldLayout|null $fieldLayout The field layout used by this element
+ * @property array $fieldParamNamespace The namespace used by custom field params on the request
+ *
  * @internal
  */
 trait HasCustomFields
@@ -35,6 +41,16 @@ trait HasCustomFields
      * @since 3.5.0
      */
     public const EVENT_REGISTER_FIELD_LAYOUTS = 'registerFieldLayouts';
+
+    /**
+     * @var int|null The element’s field layout ID
+     */
+    public ?int $fieldLayoutId = null;
+
+    /**
+     * @var bool Whether the element is still awaiting its custom field values
+     */
+    public bool $awaitingFieldValues = false;
 
     private ?array $_outdatedFields = null;
