+++ src/Element/Concerns/Localizable.php
@@ -7,6 +7,7 @@ namespace CraftCms\Cms\Element\Concerns;
 use craft\base\ElementInterface;
 use craft\base\NestedElementInterface;
 use craft\helpers\ElementHelper;
+use craft\web\twig\AllowedInSandbox;
 use CraftCms\Cms\Element\ElementCollection;
 use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
 use CraftCms\Cms\Element\Queries\ElementQuery;
@@ -15,12 +16,77 @@ use CraftCms\Cms\Site\Data\Site;
 use CraftCms\Cms\Support\Facades\Sites;
 use yii\base\InvalidConfigException;
 
+/**
+ * Localizable provides localization functionality for elements.
+ *
+ * This trait contains methods for determining site-specific properties and querying localized versions of elements.
+ *
+ * @property int[]|array $supportedSites The sites this element is associated with
+ * @property Site $site Site the element is associated with
+ *
+ * @internal
+ */
 trait Localizable
 {
     /**
      * @see getIsCrossSiteCopyable()
      */
     private bool $isCrossSiteCopyable;
+
+    /**
+     * @var int|null The site ID the element is associated with
+     */
+    #[AllowedInSandbox]
+    public ?int $siteId = null;
+
+    /**
+     * @var bool Whether the element is being saved in the context of propagating another site's version of the element.
+     */
+    public bool $propagating = false;
+
+    /**
+     * @var ElementInterface|null The element that this element is being propagated from.
+     *
+     * @since 5.0.0
+     */
+    public ?ElementInterface $propagatingFrom = null;
+
+    /**
+     * @var bool Whether all element attributes should be propagated across all its supported sites, even if that means
+     *           overwriting existing site-specific values.
+     *
+     * @since 3.2.0
+     */
+    public bool $propagateAll = false;
+
+    /**
+     * @var bool Whether all required element attributes should be propagated across all its supported sites, but only if otherwise
+     *           they wouldn’t validate.
+     *
+     * @since 5.9.0
+     */
+    public bool $propagateRequired = false;
+
+    /**
+     * @var int[] The site IDs that the element was just propagated to for the first time.
+     *
+     * @since 3.2.9
+     */
+    public array $newSiteIds = [];
+
+    /**
+     * @var bool Whether the element is being saved to the current site for the first time.
+     *
+     * @since 3.7.15
+     */
+    public bool $isNewForSite = false;
+
+    /**
+     * @var bool Whether this is for a newly-created site.
+     *
+     * @since 5.6.10
+     */
+    public bool $isNewSite = false;
 
     /**
      * {@inheritdoc}
